Se implementaron el login y registro definido, api keys de OpenAI y Gemini, implementados con el proxy y comunicación con el dashboard

+++ backend/app/schemas/dashboard.py
@@ -1,14 +1,13 @@
 from datetime import date, datetime
-from decimal import Decimal
 from typing import Optional
 
 from pydantic import BaseModel
 
 
 class SummaryResponse(BaseModel):
-    total_spend_usd: Decimal
+    total_spend_usd: float
     request_count: int
-    avg_cost_per_request: Decimal
+    avg_cost_per_request: float
     total_tokens: int
     # Change vs previous period (same duration)
     spend_change_pct: Optional[float]    # None when no previous data
@@ -17,14 +16,14 @@ class SummaryResponse(BaseModel):
 
 class DailySpend(BaseModel):
     date: date
-    spend_usd: Decimal
+    spend_usd: float
     request_count: int
 
 
 class SpendByModel(BaseModel):
     model_name: str
     display_name: str
-    spend_usd: Decimal
+    spend_usd: float
     request_count: int
     total_tokens: int
     pct_of_total: float    # 0–100
@@ -36,7 +35,7 @@ class ActivityRow(BaseModel):
     display_name: str
     tokens_input: int
     tokens_output: int
-    cost_usd: Decimal
+    cost_usd: float
     latency_ms: int
     is_streaming: bool
     created_at: datetime
